Move common functionality to common location
