+++ extension/src/.common/hooks/use-command.ts
