+++ extension/src/.common/types/preferences.ts
