+++ extension/src/.common/types/settings.ts
